+++ codes/code_providers_test.go
@@ -1,19 +1,24 @@
 package codes
 
 import (
+	"encoding/json"
 	"testing"
 
-	. "github.com/smartystreets/goconvey/convey"
 	"gitlab.com/tokend/go/xdr"
 )
 
 func TestCodeProviders(t *testing.T) {
-	Convey("Code providers", t, func() {
-		for _, opType := range xdr.OperationTypeAll {
-			_, ok := codeProviders[opType]
-			if !ok {
-				t.Errorf("Failed to get code provider for %s", opType.ShortString())
-			}
-		}
-	})
+	rawTx := "AAAAAD3w0GHAzCqJfiKeMESG6h6zyi4voJv9i5aDBOUjd4osAAAAAAAAAAAAAAAAAAAAAAAAAABcepj9AAAAAAAAAAEAAAAAAAAAEgAAAAAAAAAO8d+oN6I3bsEziNvff3+lIB6iyBdlHoIXD4RdOo44kL8AAAADAAAAAwAAAAhhc2RhIHNkYQAAAAAAAAAAAAAAAnt9AAAAAAAAAAAAAAAAAAAAAAABI3eKLAAAAEDOGMFv7GM/2cWqtFJyaFSqoIHMTzLh7pkpKxc2gLEoN+JlEhYwyo0X8YhDLWRIxooBhOUlatoD46jmpn9QDEUC"
+	var meta xdr.TransactionEnvelope
+	err := xdr.SafeUnmarshalBase64(rawTx, &meta)
+	if err != nil {
+		panic(err)
+	}
+
+	asJSON, err := json.Marshal(&meta)
+	if err != nil {
+		panic(err)
+	}
+
+	println(string(asJSON))
 }