+++ web_v2/handlers/main.go
@@ -2,6 +2,7 @@ package handlers
 
 import (
 	"net/http"
+	"strings"
 
 	"gitlab.com/distributed_lab/ape"
 	"gitlab.com/distributed_lab/ape/problems"
@@ -15,11 +16,16 @@ import (
 func isAllowed(r *http.Request, w http.ResponseWriter, dataOwners ...string) bool {
 	constraints := make([]doorman.SignerConstraint, 0, len(dataOwners))
 	for _, dataOwner := range dataOwners {
+		// invalid account address will make doorman return 401 w/o considering other constraints
+		if dataOwner == "" {
+			continue
+		}
 		constraints = append(constraints, doorman.SignerOf(dataOwner))
 	}
 	constraints = append(constraints, doorman.SignerOf(ctx.CoreInfo(r).AdminAccountID))
 
 	err := ctx.Doorman(r, constraints...)
+
 	switch err {
 	case nil:
 		return true
@@ -29,6 +35,8 @@ func isAllowed(r *http.Request, w http.ResponseWriter, dataOwners ...string) boo
 		notAllowed := problems.NotAllowed()
 		notAllowed.Meta = &map[string]interface{}{
 			"cause": err.Error(),
+			// probably useful for debug
+			"checked_against": strings.Join(append(dataOwners, ctx.CoreInfo(r).AdminAccountID), ", "),
 		}
 		ape.RenderErr(w, notAllowed)
 		return false