+++ web_v2/requests/get_request_list_base.go
@@ -1,9 +1,8 @@
 package requests
 
 import (
-	"net/http"
-
 	"gitlab.com/tokend/horizon/db2"
+	"net/http"
 )
 
 type GetRequestsBase struct {
@@ -36,6 +35,9 @@ func NewGetRequestsBase(
 		supportedFilters:  mergedFilters,
 		supportedIncludes: mergedIncludes,
 	})
+	if err != nil {
+		return nil, err
+	}
 
 	pageParams, err := b.getCursorBasedPageParams()
 	if err != nil {